feat: Align with RVX ([v5.7.1](https://github.com/inotia00/revanced-patches/releases/tag/v5.7.1))
Close https://github.com/anddea/revanced-patches/issues/1077
Close https://github.com/anddea/revanced-patches/issues/1090
Close https://github.com/anddea/revanced-patches/issues/1095
Close https://github.com/anddea/revanced-patches/issues/1095
Close https://github.com/anddea/revanced-patches/issues/1111
Close https://github.com/anddea/revanced-patches/issues/1124
Close https://github.com/anddea/revanced-patches/issues/1125
Close https://github.com/anddea/revanced-patches/issues/1135
Close https://github.com/anddea/revanced-patches/issues/1137

+++ extensions/shared/src/main/java/app/revanced/extension/youtube/swipecontrols/misc/ScrollDistanceHelper.kt
@@ -10,7 +10,7 @@ import kotlin.math.sign
  * @param callback callback function for when unit distance is reached
  */
 class ScrollDistanceHelper(
-    private val unitDistance: Double,
+    private val unitDistance: Int,
     private val callback: (oldDistance: Double, newDistance: Double, direction: Int) -> Unit,
 ) {
